refactor: move ProviderType to shared types and update usages
