+++ ui/components/providers/workflow/forms/test-connection-form.tsx
@@ -18,9 +18,9 @@ import { getTask } from "@/actions/task/tasks";
 import { CheckIcon, RocketIcon } from "@/components/icons";
 import { useToast } from "@/components/ui";
 import { CustomButton } from "@/components/ui/custom";
-import { ProviderType } from "@/components/ui/entities";
 import { Form } from "@/components/ui/form";
 import { checkTaskStatus } from "@/lib/helper";
+import { ProviderType } from "@/types";
 import { ApiError, testConnectionFormSchema } from "@/types";
 
 import { ProviderInfo } from "../..";
